Fix :    - correction du nom du BlocRegles.vue    - correction de l'algorithme des filtres de la dataTable du BlocRegles.vue    - correction de la taille de l'affichage du contenu du BlocHistorique.vue    - correction du positionnement des éléments d'affichage et de suppression des filtres du BlocRegles.vue
